# DM/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /users:
        get:
            tags:
                - UserService
            operationId: UserService_GetAllUser
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetAllUserRespone'
        post:
            tags:
                - UserService
            operationId: UserService_SignUpUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.SignUpUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.SignUpUserRespone'
    /users/page:
        get:
            tags:
                - UserService
            operationId: UserService_GetUserByPage
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.GetUsertByPageRespone'
    /users/refresh_access_token:
        post:
            tags:
                - UserService
            operationId: UserService_RefreshAccessToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.RefreshAccessTokenRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.RefreshAccessTokenRespone'
    /users/signin:
        post:
            tags:
                - UserService
            operationId: UserService_SignInUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/helloworld.v1.SignInUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.SignInUserRespone'
    /users/{id}:
        delete:
            tags:
                - UserService
            operationId: UserService_DeleteUser
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.DeleteUserRespone'
components:
    schemas:
...
        helloworld.v1.DeleteUserRespone:
            type: object
            properties: {}
...
        helloworld.v1.GetAllUserRespone:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.User'
...
        helloworld.v1.GetUsertByPageRespone:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/helloworld.v1.User'
...
        helloworld.v1.RefreshAccessTokenRequest:
            type: object
            properties:
                refreshToken:
                    type: string
        helloworld.v1.RefreshAccessTokenRespone:
            type: object
            properties:
                accessToken:
                    type: string
        helloworld.v1.SignInUserRequest:
            type: object
            properties:
                phone:
                    type: string
                password:
                    type: string
                role:
                    type: string
        helloworld.v1.SignInUserRespone:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
        helloworld.v1.SignUpUserRequest:
            type: object
            properties:
                phone:
                    type: string
                password:
                    type: string
                confirmpass:
                    type: string
                isEmployee:
                    type: boolean
                role:
                    type: string
        helloworld.v1.SignUpUserRespone:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                phone:
                    type: string
                password:
                    type: string
                isEmployee:
                    type: boolean
                role:
                    type: string
...
        helloworld.v1.User:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                phone:
                    type: string
                password:
                    type: string
                confirmpass:
                    type: string
                isEmployee:
                    type: boolean
                role:
                    type: string
                accessToken:
                    type: string
                refreshToken:
                    type: string
...